initial: 1.15.2 yarn mapping

+++ src/net/minecraft/block/MyceliumBlock.java
@@ -28,7 +28,7 @@ extends SpreadableBlock {
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         super.randomDisplayTick(state, world, pos, random);
         if (random.nextInt(10) == 0) {
-            world.addParticle(ParticleTypes.MYCELIUM, (float)pos.getX() + random.nextFloat(), (float)pos.getY() + 1.1f, (float)pos.getZ() + random.nextFloat(), 0.0, 0.0, 0.0);
+            world.addParticle(ParticleTypes.MYCELIUM, (double)pos.getX() + (double)random.nextFloat(), (double)pos.getY() + 1.1, (double)pos.getZ() + (double)random.nextFloat(), 0.0, 0.0, 0.0);
         }
     }
 }
